pyproject.toml
[project]
name = "scrolling-video-panorama"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "av>=14.4.0",
    "ipython>=9.3.0",
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "torch>=2.7.1",
    "torchvideo>=0.0.0",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "typer>=0.16.0",
]
